+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1
 
 - solve issue when exporting in PNG
+- solve device issue when changing sampling rate (#17)
 
 
 ## v0.1.0 - 2023-10-17